click/requirements.txt
# KFC Project + API Click & Collect
requests
flask
python-dotenv
psycopg2-binary
